backend/requirements.txt
python-binance==1.0.19
typer>=0.12.0
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
httpx==0.26.0
